// src/dto/synchronization.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface definitions {
  AuthorizationInformation: {
    isUserAuthorized: boolean;
    _links: definitions['authorization-information-links'];
  };
  'authorization-information-links': {
    authorizationUrl: definitions['Link'];
  };
  'connections-prefer-return-minimal': {
    connections: definitions['connection-summary'][];
    _links: definitions['Links'];
  };
  'connections-prefer-return-representation': {
    connections: definitions['Connection'][];
    _links: definitions['Links'];
  };
  Connection: {
    id: string;
    displayName: string;
    iModelId: string;
    _links: definitions['connection-links'];
  };
  'connection-summary': {
    id: string;
    displayName: string;
  };
  StorageConnection: {
    id: string;
    displayName: string;
    iModelId: string;
    _links: definitions['connection-links'];
  };
  'storage-connection-summary': {
    id: string;
    displayName: string;
  };
  'storage-connection-create': {
    displayName: string;
    iModelId: string;
    sourceFiles: definitions['storage-file-create'][];
  };
  'runs-prefer-return-minimal': {
    runs: definitions['run-summary'][];
  };
  'runs-prefer-return-representation': {
    runs: definitions['Run'][];
  };
  Run: {
    id: string;
    connectionId: string;
    startDateTime: string;
    endDateTime: string;
    phase: string;
    state: string;
    result: string;
    jobs: definitions['Job'][];
  };
  'run-summary': {
    id: string;
    state: string;
    result: string;
  };
  Job: {
    id: string;
    startDateTime: string;
    endDateTime: string;
    state: string;
    result: string;
    connectorType: definitions['connector-type'];
    tasks: definitions['Task'][];
  };
  Task: {
    id: string;
    startDateTime: string;
    endDateTime: string;
    retryAttempts: number;
    storageFileId: string;
    state: string;
    result: string;
  };
  'storage-file-prefer-return-minimal': {
    sourceFiles: definitions['StorageFile (summary)'][];
  };
  'storage-file-prefer-return-representation': {
    sourceFiles: definitions['StorageFile'][];
  };
  StorageFile: {
    id: string;
    storageFileId: string;
    connectorType: definitions['connector-type'];
    lastKnownFileName: string;
    _links: definitions['file-links'];
  };
  'storage-file-summary': {
    id: string;
    storageFileId: string;
  };
  'storage-file-create': {
    storageFileId: string;
    connectorType: definitions['connector-type'];
  };
  'connection-links': {
    iModel: definitions['Link'];
    project: definitions['Link'];
    lastRun?: definitions['Link'];
  };
  'file-links': {
    file: definitions['Link'];
  };
  Link: {
    href: string;
  };
  Links: {
    self: definitions['Link'];
    next: definitions['Link'];
    prev: definitions['Link'];
  };
  /** One of 'NotSet', 'AFFINITYCHECKER, 'MSTN', 'AVEVAPID', 'AUTOPLANT', 'REVIT', 'DWG', 'SPXREVIEW', 'IFC', 'SPPID', 'AVEVADIAGRAMS', 'CIVIL', 'CIVIL3D', 'GEOSPATIAL', 'OBD', 'OPENTOWER', 'SHELLEDWCSV', 'NWD'. Name of the Connector that is used for synchronization. */
  'connector-type':
    | 'NotSet'
    | 'AFFINITYCHECKER'
    | 'MSTN'
    | 'AVEVAPID'
    | 'AUTOPLANT'
    | 'REVIT'
    | 'DWG'
    | 'SPXREVIEW'
    | 'IFC'
    | 'SPPID'
    | 'AVEVADIAGRAMS'
    | 'CIVIL'
    | 'CIVIL3D'
    | 'GEOSPATIAL'
    | 'OBD'
    | 'OPENTOWER'
    | 'SHELLEDWCSV'
    | 'NWD';
}